Add Node class
There's a new Node class with some basic methods to interface with a
plain object. The object is protected except for an `update` method
which mutates it, but soon it'll be replaced with `merge`, and maybe an
immutable method that does something similar... dunno yet.
JSDoc has been added, as well as a beautiful theme for it and a
configuration file.
All tests passing.