Add uuids to nodes, add uuid method
The Node class has a new static method `uuid` which generates random
strings recursively. Every new node created will have a uid property
that is unique to that node (soon to be configurable). The `Node.uid`
static method is completely editable, accepting a character set and a
length option, but defaulting to alphanumeric symbols and 24 character
strings. All tests passing.