Fix Graph#merge, add entry file.
The graph `.merge` method was assuming nodes within a graph were
unformatted. This made syncing with other peers more difficult, since
graphs are the exchange format, and there was no easy way to import a
serialized graph into another. All tests passing and a new one added.
A new file has been added that exports both Graph and Node as
properties. It's the main entry point to the package.