Drop core support for aggregate nodes
It was an idea to enable secondary indices directly via the graph
interface. However, since this library is used as a lower level data
structure (for example the Mytosis database), any aggregate plugins
should be left to a higher level, where more app details are known and
less assumptions need to be imposed.