Replace Node#clone with Node#new
The clone method couldn't be used in some circumstances because it
copied the node's properties, which is very difficult to undo. Instead,
clone has been dropped in favor of Node#new(), which creates a new
instance of the same type, under the same configuration as the original
(e.g., uid and any subclass arguments).
Changelog updated.