Rename Node#read to Node#value
The .read method made too much sense, so I changed it.

Kidding.

The name .read had to be freed up for subclasses to use, specifically
in the Mytosis project. The danger and tedium of overloading a method
in a subclass was more complex than simply renaming the method.
Renamed to .value for its parallel to the .state method.