Add `Entity` base class
Defines common methods over the data structure, but does not specify
merge implementations. This is left up to subclasses. After attempting
to implement a linked list over `Node`, I determined the class went too
deep and changing core methods at that level broke a lot of assumptions.
This new class will be the base of both `Node` and `List`.