Move delta logic into Entity class
It seems there are two parts to a merge:
- Figure out what changed
- Apply and emit those changes

Since `List` will need to know the merge deltas _before_ it mutates,
mutation can't be part of the superclass. By separating delta logic and
mutation, `List` can safely share the code and `Node` doesn't need so
many lines in the `merge()` function. Win win!